requirements.txt
django>=2.0
python-telegram-bot>=10.1.0